change: 貸出中に削除できない機能を追加

+++ src/main/java/jp/co/seattle/library/controller/DeleteBookController.java
@@ -1,9 +1,5 @@
 package jp.co.seattle.library.controller;
 
-import java.util.Locale;
-
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
 import org.springframework.transaction.annotation.Transactional;
@@ -13,16 +9,19 @@ import org.springframework.web.bind.annotation.RequestMethod;
 import org.springframework.web.bind.annotation.RequestParam;
 
 import jp.co.seattle.library.service.BooksService;
+import jp.co.seattle.library.service.RentalsService;
 
 /**
  * 削除コントローラー
  */
 @Controller //APIの入り口
 public class DeleteBookController {
-    final static Logger logger = LoggerFactory.getLogger(DeleteBookController.class);
     
     @Autowired
     private BooksService booksService;
+    
+    @Autowired
+    private RentalsService rentalsService;
 
     /**
      * 対象書籍を削除する
@@ -35,11 +34,17 @@ public class DeleteBookController {
     @Transactional
     @RequestMapping(value = "/deleteBook", method = RequestMethod.POST)
     public String deleteBook(
-            Locale locale,
             @RequestParam("bookId") Integer bookId,
             Model model) {
-        logger.info("Welcome delete! The client locale is {}.", locale);
-        booksService.deleteBook(bookId);
+        
+		if(rentalsService.getRentInfo(bookId) == 0) {
+			booksService.deleteBook(bookId);
+		} else {
+			model.addAttribute("rentErrorMessage", "貸出しされているため削除できません。");
+			model.addAttribute("bookDetailsInfo", booksService.getBookInfo(bookId));
+			return "details";
+		}
+		
         model.addAttribute("bookList", booksService.getBookList());
         return "home";
